@@ -1,8 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+      <TargetFramework>netcoreapp3.1</TargetFramework>
         <UserSecretsId>User-Secret-Academie.PawnShop</UserSecretsId>
+      <AddRazorSupportForMvc>true</AddRazorSupportForMvc>
     </PropertyGroup>
 
     <ItemGroup>
@@ -14,17 +15,19 @@
 
     <ItemGroup>
         <PackageReference Include="FluentEmail.Razor" Version="2.7.0" />
-        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.7.1" />
+        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.13.0-beta2" />
+        <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.1" />
         <PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="2.2.0" />
-        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.3" />
-        <PackageReference Include="Serilog" Version="2.8.0" />
-        <PackageReference Include="Serilog.Extensions.Logging" Version="3.0.1" />
+        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.1" />
+        <PackageReference Include="Serilog" Version="2.9.1-dev-01154" />
+        <PackageReference Include="Serilog.Extensions.Logging" Version="3.0.2-dev-10265" />
         <PackageReference Include="Serilog.Sinks.ColoredConsole" Version="3.0.1" />
-        <PackageReference Include="Serilog.Sinks.RollingFile" Version="3.3.0" />
+        <PackageReference Include="Serilog.Sinks.RollingFile" Version="3.3.1-dev-00771" />
         <PackageReference Include="Serilog.Sinks.Sentry.AspNetCore" Version="2.4.2" />
         <PackageReference Include="Serilog.Sinks.SyslogServer" Version="1.0.10" />
-        <PackageReference Include="Spk.Common.Helpers" Version="1.8.1" />
-        <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.6" />
+        <PackageReference Include="Spk.Common.Helpers" Version="1.11.1" />
+        <PackageReference Include="Microsoft.AspNetCore.App" Version="3.0" />
     </ItemGroup>
 
     <ItemGroup>
